requirements.txt
email_validator==2.1.0.post1
flask
flask-login
flask-sqlalchemy
flask-wtf
pandas
google-api-python-client
python-dotenv 
google-generativeai
networkx
numpy
plotly
sqlalchemy
werkzeug
wtforms
youtube-transcript-api
authlib
requests
python-jose
tavily-python
atproto
groq
twilio